+++ src/README.md
@@ -1,5 +1,7 @@
-# Project Name Source Code
+# Park'd Source Code
 
 The folders and files for this project are as follows:
 
-...
+- [park_d_model_service](https://github.com/parkd-app/park-d/tree/main/src/park_d_model_service) - Model
+- [park-d-api](https://github.com/parkd-app/park-d/tree/main/src/park-d-api) - API
+- [park_d_front-end](https://github.com/parkd-app/park-d/tree/main/src/park-d-front-end) - Frontend
