--- src/pard-d-api/Constants/constants.py
@@ -1,36 +1,0 @@
-PARKING_INFO_PATH = (
-    "C:\\Users\\g7543\\OneDrive\\桌面\\"
-    "Capstone_project\\park-d\\"
-    "src\\park_d_model_service\\gaussian_model\\"
-    "parking_lot_model_v1\\parking_info.txt"
-)
-
-ROOT =  ( "C:\\Users\\g7543\\OneDrive\\桌面\\"
-    "Capstone_project\\park-d\\"
-    "src\\park_d_model_service\\"
-)
-
-PARKING_INFO_PATH_BIRD = (
-   ROOT + "\\Bird_view model\\parking_info.txt"
-)
-
-VIDEO_PATH_BIRD = (
-    "C:\\Users\\g7543\\OneDrive\\" "new training footage\\train2_bird.mp4"
-)
-
-VIDEO_PATH_SIDE = (
-    "C:\\Users\\g7543\\OneDrive\\" "new training footage\\train1_side.mov"
-)
-
-
-IMAGE_PATH = (
-    "C:\\Users\\g7543\\OneDrive\\Capstone_project\\"
-    "park-d\\src\\park_d_model_service\\gaussian_model\\"
-    "parking_lot_model_v1\\images"
-)
-
-model_side_path = (
-    "C:\\Users\\g7543\\OneDrive\\Cpastone_final\\"
-    "park-d\\src\\park_d_model_service\\gaussian_model\\"
-    "parking_lot_model_v1\\main.py"
-)
