+++ src/park-d-api/static_resources/bird.png
