+++ src/park-d-api/static_resources/coordinates_bird
