+++ src/park-d-api/static_resources/parking_lot_2.png
